Add files via upload
Problem Statement
As a beginner in data analytics, the objective of this assignment is to analyze the UK road accident dataset and gain insights into factors affecting road safety and accident severity. By examining the dataset, we aim to explore various aspects of road accidents and address the following questions. By addressing these questions and conducting exploratory analysis on the UK road accident dataset, we aim to gain insights into road safety and accident prevention.


Use Advanced Power BI Visualisations and Filters.

What are total Casualties, average speed limit?

Which police force has the highest number of reported accidents?

Which districts have the more than 1000 number of reported accidents?

Which vehicle types are involved in the highest number of accidents?

What is the relationship between the number of vehicles involved in an accident and the average of casualties?

What is the trend of accidents over date (year, quarter, months and even day) in the UK?

Are there certain junction types that are more commonly associated with accidents?

Which districts are the most affected in terms of accidents?

How do road surface conditions affect accident severity?

Add a slicer for accident severity and district to filter the entire report page.